Add infrastructure layer
