--- src/homeworks-api/src/AspnetRun.Infrastructure/Data/AspnetRunContext.cs
@@ -1,44 +1,0 @@
-﻿using AspnetRun.Core.Entities;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace AspnetRun.Infrastructure.Data
-{
-    public class AspnetRunContext : DbContext
-    {
-        public AspnetRunContext(DbContextOptions options) : base(options)
-        {
-        }
-
-        public DbSet<Product> Products { get; set; }
-        public DbSet<Category> Categories { get; set; }
-
-        protected override void OnModelCreating(ModelBuilder builder)
-        {
-            builder.Entity<Product>(ConfigureProduct);
-            builder.Entity<Category>(ConfigureCategory);
-        }
-
-        private void ConfigureProduct(EntityTypeBuilder<Product> builder)
-        {
-            builder.ToTable("Product");
-
-            builder.HasKey(ci => ci.Id);         
-
-            builder.Property(cb => cb.ProductName)
-                .IsRequired()
-                .HasMaxLength(100);
-        }
-
-        private void ConfigureCategory(EntityTypeBuilder<Category> builder)
-        {
-            builder.ToTable("Category");
-
-            builder.HasKey(ci => ci.Id);            
-
-            builder.Property(cb => cb.CategoryName)
-                .IsRequired()
-                .HasMaxLength(100);
-        }
-    }
-}
