--- src/homeworks-api/src/AspnetRun.Infrastructure/Data/AspnetRunContextSeed.cs
@@ -1,67 +1,0 @@
-﻿using AspnetRun.Core.Entities;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace AspnetRun.Infrastructure.Data
-{
-    public class AspnetRunContextSeed
-    {
-        public static async Task SeedAsync(AspnetRunContext aspnetrunContext, ILoggerFactory loggerFactory, int? retry = 0)
-        {
-            int retryForAvailability = retry.Value;
-
-            try
-            {
-                // TODO: Only run this if using a real database
-                // aspnetrunContext.Database.Migrate();
-                // aspnetrunContext.Database.EnsureCreated();
-
-                if (!aspnetrunContext.Categories.Any())
-                {
-                    aspnetrunContext.Categories.AddRange(GetPreconfiguredCategories());
-                    await aspnetrunContext.SaveChangesAsync();
-                }
-
-                if (!aspnetrunContext.Products.Any())
-                {
-                    aspnetrunContext.Products.AddRange(GetPreconfiguredProducts());
-                    await aspnetrunContext.SaveChangesAsync();
-                }
-            }
-            catch (Exception exception)
-            {
-                if (retryForAvailability < 10)
-                {
-                    retryForAvailability++;
-                    var log = loggerFactory.CreateLogger<AspnetRunContextSeed>();
-                    log.LogError(exception.Message);
-                    await SeedAsync(aspnetrunContext, loggerFactory, retryForAvailability);
-                }
-                throw;
-            }
-        }
-
-        private static IEnumerable<Category> GetPreconfiguredCategories()
-        {
-            return new List<Category>()
-            {
-                new Category() { CategoryName = "Phone"},
-                new Category() { CategoryName = "TV"}
-            };
-        }
-
-        private static IEnumerable<Product> GetPreconfiguredProducts()
-        {
-            return new List<Product>()
-            {
-                new Product() { ProductName = "IPhone", CategoryId = 1 , UnitPrice = 19.5M , UnitsInStock = 10, QuantityPerUnit = "2", UnitsOnOrder = 1, ReorderLevel = 1, Discontinued = false },
-                new Product() { ProductName = "Samsung", CategoryId = 1 , UnitPrice = 33.5M , UnitsInStock = 10, QuantityPerUnit = "2", UnitsOnOrder = 1, ReorderLevel = 1, Discontinued = false },
-                new Product() { ProductName = "LG TV", CategoryId = 2 , UnitPrice = 33.5M , UnitsInStock = 10, QuantityPerUnit = "2", UnitsOnOrder = 1, ReorderLevel = 1, Discontinued = false }
-            };
-        }
-    }
-}
