--- src/homeworks-api/src/AspnetRun.Infrastructure/Exceptions/InfrastructureException.cs
@@ -1,17 +1,0 @@
-﻿using System;
-
-namespace AspnetRun.Infrastructure.Exceptions
-{
-    public class InfrastructureException : Exception
-    {
-        internal InfrastructureException(string businessMessage)
-               : base(businessMessage)
-        {
-        }
-
-        internal InfrastructureException(string message, Exception innerException)
-            : base(message, innerException)
-        {
-        }
-    }
-}
