--- src/homeworks-api/src/AspnetRun.Infrastructure/Logging/LoggerAdapter.cs
@@ -1,25 +1,0 @@
-﻿using AspnetRun.Core.Interfaces;
-using Microsoft.Extensions.Logging;
-
-namespace AspnetRun.Infrastructure.Logging
-{
-    public class LoggerAdapter<T> : IAppLogger<T>
-    {
-        private readonly ILogger<T> _logger;
-
-        public LoggerAdapter(ILoggerFactory loggerFactory)
-        {
-            _logger = loggerFactory.CreateLogger<T>();
-        }
-
-        public void LogWarning(string message, params object[] args)
-        {
-            _logger.LogWarning(message, args);
-        }
-
-        public void LogInformation(string message, params object[] args)
-        {
-            _logger.LogInformation(message, args);
-        }      
-    }
-}
