--- src/homeworks-api/src/AspnetRun.Infrastructure/Repository/Base/SpecificationEvaluator.cs
@@ -1,48 +1,0 @@
-﻿using AspnetRun.Core.Entities;
-using AspnetRun.Core.Entities.Base;
-using AspnetRun.Core.Specifications.Base;
-using Microsoft.EntityFrameworkCore;
-using System.Linq;
-
-namespace AspnetRun.Infrastructure.Repository.Base
-{
-    public class SpecificationEvaluator<T> where T : Entity
-    {
-        public static IQueryable<T> GetQuery(IQueryable<T> inputQuery, ISpecification<T> specification)
-        {
-            var query = inputQuery;
-
-            // modify the IQueryable using the specification's criteria expression
-            if (specification.Criteria != null)
-            {
-                query = query.Where(specification.Criteria);
-            }
-
-            // Includes all expression-based includes
-            query = specification.Includes.Aggregate(query,
-                                    (current, include) => current.Include(include));
-
-            // Include any string-based include statements
-            query = specification.IncludeStrings.Aggregate(query,
-                                    (current, include) => current.Include(include));
-
-            // Apply ordering if expressions are set
-            if (specification.OrderBy != null)
-            {
-                query = query.OrderBy(specification.OrderBy);
-            }
-            else if (specification.OrderByDescending != null)
-            {
-                query = query.OrderByDescending(specification.OrderByDescending);
-            }
-
-            // Apply paging if enabled
-            if (specification.isPagingEnabled)
-            {
-                query = query.Skip(specification.Skip)
-                             .Take(specification.Take);
-            }
-            return query;
-        }
-    }
-}
