--- src/homeworks-api/src/AspnetRun.Infrastructure/Repository/CategoryRepository.cs
@@ -1,24 +1,0 @@
-﻿using AspnetRun.Core.Entities;
-using AspnetRun.Core.Repositories;
-using AspnetRun.Core.Specifications;
-using AspnetRun.Infrastructure.Data;
-using AspnetRun.Infrastructure.Repository.Base;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace AspnetRun.Infrastructure.Repository
-{
-    public class CategoryRepository : Repository<Category>, ICategoryRepository
-    {
-        public CategoryRepository(AspnetRunContext dbContext) : base(dbContext)
-        {            
-        }
-
-        public async Task<Category> GetCategoryWithProductsAsync(int categoryId)
-        {            
-            var spec = new CategoryWithProductsSpecification(categoryId);
-            var category = (await GetAsync(spec)).FirstOrDefault();
-            return category;
-        }
-    }
-}
