--- src/homeworks-api/src/AspnetRun.Infrastructure/Services/EmailSender.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace AspnetRun.Infrastructure.Services
-{
-    // This class is used by the application to send email for account confirmation and password reset.
-    // For more details see https://go.microsoft.com/fwlink/?LinkID=532713
-    public class EmailSender // : IEmailSender
-    {
-        public Task SendEmailAsync(string email, string subject, string message)
-        {
-            // TODO: Wire this up to actual email sending logic via SendGrid, local SMTP, etc.
-            return Task.CompletedTask;
-        }
-    }
-}
