+++ src/homeworks-api/src/EKids.Chatbot.Homeworks.Infrastructure/EKids.Chatbot.Homeworks.Infrastructure.csproj
@@ -1,7 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
@@ -19,7 +20,12 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\AspnetRun.Core\AspnetRun.Core.csproj" />
+    <ProjectReference Include="..\EKids.Chatbot.Homeworks.Core\EKids.Chatbot.Homeworks.Core.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Migrations\" />
+    <Folder Include="Services\" />
   </ItemGroup>
 
 </Project>
