+++ src/homeworks-api/src/EKids.Chatbot.Homeworks.Infrastructure/Exceptions/InfrastructureException.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace EKids.Chatbot.Homeworks.Infrastructure.Exceptions;
+
+public class InfrastructureException : Exception
+{
+    internal InfrastructureException(string businessMessage)
+           : base(businessMessage)
+    {
+    }
+
+    internal InfrastructureException(string message, Exception innerException)
+        : base(message, innerException)
+    {
+    }
+}
