+++ src/homeworks-api/src/EKids.Chatbot.Homeworks.Infrastructure/Repository/Base/Repository.cs
@@ -1,0 +1,100 @@
+﻿using EKids.Chatbot.Homeworks.Core.Entities.Base;
+using EKids.Chatbot.Homeworks.Core.Repositories.Base;
+using EKids.Chatbot.Homeworks.Core.Specifications.Base;
+using EKids.Chatbot.Homeworks.Infrastructure.Data;
+using Microsoft.EntityFrameworkCore;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Threading.Tasks;
+
+namespace EKids.Chatbot.Homeworks.Infrastructure.Repository.Base;
+
+public class Repository<T>(HomeworkContext dbContext) : IRepository<T> where T : Entity
+{
+    protected readonly HomeworkContext _dbContext = dbContext ?? throw new ArgumentNullException(nameof(dbContext));
+
+    public async Task<IReadOnlyList<T>> GetAllAsync()
+    {
+        return await _dbContext.Set<T>().ToListAsync();
+    }
+
+    public async Task<IReadOnlyList<T>> GetAsync(ISpecification<T> spec)
+    {
+        return await ApplySpecification(spec).ToListAsync();
+    }
+
+    public async Task<int> CountAsync(ISpecification<T> spec)
+    {
+        return await ApplySpecification(spec).CountAsync();
+    }
+
+    private IQueryable<T> ApplySpecification(ISpecification<T> spec)
+    {
+        return SpecificationEvaluator<T>.GetQuery(_dbContext.Set<T>().AsQueryable(), spec);
+    }
+
+    public async Task<IReadOnlyList<T>> GetAsync(Expression<Func<T, bool>> predicate)
+    {
+        return await _dbContext.Set<T>().Where(predicate).ToListAsync();
+    }
+
+    public async Task<IReadOnlyList<T>> GetAsync(Expression<Func<T, bool>>? predicate = null, Func<IQueryable<T>, IOrderedQueryable<T>>? orderBy = null, string? includeString = null, bool disableTracking = true)
+    {
+        IQueryable<T> query = _dbContext.Set<T>();
+        if (disableTracking) query = query.AsNoTracking();
+
+        if (!string.IsNullOrWhiteSpace(includeString)) query = query.Include(includeString);
+
+        if (predicate != null) query = query.Where(predicate);
+
+        if (orderBy != null)
+        {
+            return await orderBy(query).ToListAsync();
+        }
+
+        return await query.ToListAsync();
+    }
+
+    public async Task<IReadOnlyList<T>> GetAsync(Expression<Func<T, bool>>? predicate = null, Func<IQueryable<T>, IOrderedQueryable<T>>? orderBy = null, List<Expression<Func<T, object>>>? includes = null, bool disableTracking = true)
+    {
+        IQueryable<T> query = _dbContext.Set<T>();
+        if (disableTracking) query = query.AsNoTracking();
+
+        if (includes != null) query = includes.Aggregate(query, (current, include) => current.Include(include));
+
+        if (predicate != null) query = query.Where(predicate);
+
+        if (orderBy != null)
+        {
+            return await orderBy(query).ToListAsync();
+        }
+
+        return await query.ToListAsync();
+    }
+
+    public virtual async Task<T?> GetByIdAsync(Guid id)
+    {
+        return await _dbContext.Set<T>().FindAsync(id);
+    }
+
+    public async Task<T> AddAsync(T entity)
+    {
+        _dbContext.Set<T>().Add(entity);
+        await _dbContext.SaveChangesAsync();
+        return entity;
+    }
+
+    public async Task UpdateAsync(T entity)
+    {
+        _dbContext.Entry(entity).State = EntityState.Modified;
+        await _dbContext.SaveChangesAsync();
+    }
+
+    public async Task DeleteAsync(T entity)
+    {
+        _dbContext.Set<T>().Remove(entity);
+        await _dbContext.SaveChangesAsync();
+    }
+}
