+++ src/homeworks-api/src/EKids.Chatbot.Homeworks.Infrastructure/Repository/HomeworkRepository.cs
@@ -1,0 +1,9 @@
+﻿using EKids.Chatbot.Homeworks.Core.Entities;
+using EKids.Chatbot.Homeworks.Infrastructure.Data;
+using EKids.Chatbot.Homeworks.Infrastructure.Repository.Base;
+
+namespace EKids.Chatbot.Homeworks.Infrastructure.Repository;
+
+public class HomeworkRepository(HomeworkContext dbContext) : Repository<Homework>(dbContext)
+{
+}
